pyproject.toml
[project]
name = "career-campass"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.9.1",
    "attrs==25.3.0",
    "certifi==2025.8.3",
    "charset-normalizer==3.4.3",
    "cloudinary==1.44.1",
    "colorama==0.4.6",
    "django==5.2.5",
    "django-cloudinary-storage==0.3.0",
    "django-filter==25.1",
    "djangorestframework==3.16.1",
    "djangorestframework-simplejwt==5.5.1",
    "drf-spectacular==0.28.0",
    "httpx>=0.28.1",
    "idna==3.10",
    "inflection==0.5.1",
    "iniconfig==2.1.0",
    "ipython>=9.5.0",
    "jsonschema==4.25.1",
    "jsonschema-specifications==2025.4.1",
    "markdown==3.8.2",
    "packaging==25.0",
    "pillow==11.3.0",
    "pluggy==1.6.0",
    "psycopg2-binary>=2.9.10",
    "pygments==2.19.2",
    "pyjwt==2.10.1",
    "pytest==8.4.1",
    "pytest-django==4.11.1",
    "python-dotenv>=1.1.1",
    "pyyaml==6.0.2",
    "referencing==0.36.2",
    "requests==2.32.5",
    "respx>=0.22.0",
    "rpds-py==0.27.1",
    "six==1.17.0",
    "sqlparse==0.5.3",
    "typing-extensions==4.15.0",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.5.0",
]
